improve: SoList hiding when not `dlclose`d
This commit improves the hiding of a module from SoList-based detections when it does not opt to be "dlclose"d by ReZygisk. In some scenarios, the debug map would still be abnormal despite being removed manually, so, to avoid that, we call the full "soinfo_unload". However, this would call the deconstructors of the Zygisk module, causing issues in execution due to access to unitialized variables, hence we need to trick the linker into thinking it does not have deconstructors by setting the deconstructor array size to 0, and the deconstructor function to "NULL".